Update README.md
The Grade Point Average (GPA) Calculator is a web-based programme that assists students in calculating their GPA for a given semester. The user can submit their marks and corresponding credits for each subject after inputting their name, semester number, and the number of subjects for that semester. The GPA is then calculated using a preset grading system, with specific grade points assigned for distinct mark ranges. The final GPA is calculated by adding the sum of grade point and credit products and dividing by the total number of credits. After successfully calculating their GPA, the user is rewarded with a personalised congratulatory message and an emoji.